service/gpu_service/requirements.txt
Flask
flask_cors
gunicorn
google-cloud-pubsub